requirements3.txt
--allow-external mysql-connector-python
mysql-connector-python
PyYAML
